[1265] ICPC does not drop non existant table anymore. Removed unused messages_id

+++ ch.elexis.icpc/src/ch/elexis/icpc/IcpcCode.java
@@ -45,11 +45,9 @@ public class IcpcCode extends PersistentObject implements IDiagnose {
 			"exclusion", "consider", "note", "synonyms");
 		IcpcCode ver = IcpcCode.load("ver");
 		if (!ver.exists()) {
-			try {
+			if (PersistentObject.tableExists(TABLENAME)) {
 				createOrModifyTable("DROP TABLE " + TABLENAME);
-			} catch (Exception e) {
-				// janusode
-			}
+			} 
 			createOrModifyTable(createDB);
 		}
 		VersionInfo vi = new VersionInfo(ver.get("text"));
